Improved auth dialog
