+++ client/components/login-form/styles.module.sass
@@ -25,7 +25,3 @@
     justify-content: space-between
     align-items: center
     margin: 15px 0 0
-
-    .divider
-      color: $colorTextSecond
-      margin: 0 10px
