+++ client/package.json
@@ -1,6 +1,6 @@
 {
     "name": "geometki",
-    "version": "1.1.6",
+    "version": "1.1.7",
     "private": true,
     "engines": {
         "node": ">=20.11.0"
